+++ pmdarima/arima/tests/test_arima.py
@@ -1,18 +1,18 @@
 # -*- coding: utf-8 -*-
 
+"""
+Tests of the ARIMA class
+"""
+
 import numpy as np
 import pandas as pd
 
 from pmdarima.arima import ARIMA, auto_arima, AutoARIMA
 from pmdarima.arima.arima import VALID_SCORING
-from pmdarima.arima.utils import nsdiffs
-from pmdarima.arima.warnings import ModelFitWarning
-from pmdarima.compat.pytest import pytest_error_str, pytest_warning_messages
-from pmdarima.datasets import load_lynx, load_wineind, load_heartrate, \
-    load_austres
+from pmdarima.compat.pytest import pytest_error_str
+from pmdarima.datasets import load_lynx, load_wineind, load_heartrate
 from pmdarima.utils import get_callable
 
-from os.path import abspath, dirname
 from numpy.random import RandomState
 from numpy.testing import assert_array_almost_equal, assert_almost_equal, \
     assert_allclose
@@ -30,9 +30,6 @@ import time
 rs = RandomState(42)
 y = rs.rand(25)
 
-# more interesting heart rate data (asserts we can use a series)
-hr = load_heartrate(as_series=True)
-
 # > set.seed(123)
 # > abc <- rnorm(50, 5, 1)
 abc = np.array([4.439524, 4.769823, 6.558708, 5.070508,
@@ -49,22 +46,9 @@ abc = np.array([4.439524, 4.769823, 6.558708, 5.070508,
                 6.207962, 3.876891, 4.597115, 4.533345,
                 5.779965, 4.916631])
 
-austres = load_austres()
+hr = load_heartrate(as_series=True)
 wineind = load_wineind()
 lynx = load_lynx()
-
-# Yes, m is ACTUALLY 12... but that takes a LONG time. If we set it to
-# 1, we actually get a much, much faster model fit. We can only use this
-# if we're NOT testing the output of the model, but just the functionality!
-wineind_m = 1
-
-# A random xreg for the wineind array
-wineind_xreg = rs.rand(wineind.shape[0], 2)
-
-
-def _unlink_if_exists(path):
-    if os.path.exists(path):
-        os.unlink(path)
 
 
 def test_basic_arma():
@@ -109,6 +93,27 @@ def test_basic_arma():
     _, intervals = arma.predict(n_periods=10, return_conf_int=True,
                                 alpha=0.05)
     assert_array_almost_equal(intervals, expected_intervals)
+
+
+def test_issue_30():
+    # From the issue:
+    vec = np.array([33., 44., 58., 49., 46., 98., 97.])
+
+    arm = AutoARIMA(out_of_sample_size=1, seasonal=False,
+                    suppress_warnings=True)
+    arm.fit(vec)
+
+    # This is a way to force it:
+    ARIMA(order=(0, 1, 0), out_of_sample_size=1).fit(vec)
+
+    # Want to make sure it works with exog arrays as well
+    exog = np.random.RandomState(1).rand(vec.shape[0], 2)
+    auto_arima(vec, exogenous=exog, out_of_sample_size=1,
+               seasonal=False,
+               suppress_warnings=True)
+
+    # This is a way to force it:
+    ARIMA(order=(0, 1, 0), out_of_sample_size=1).fit(vec, exogenous=exog)
 
 
 @pytest.mark.parametrize(
@@ -266,6 +271,7 @@ def test_oob_for_issue_28():
 
 
 # Test the OOB functionality for SARIMAX (Issue #28) --------------------------
+
 def test_oob_sarimax():
     xreg = rs.rand(wineind.shape[0], 2)
     fit = ARIMA(order=(1, 1, 1),
@@ -307,78 +313,33 @@ def test_oob_sarimax():
 
 
 # Test Issue #29 (d=0, cv=True) -----------------------------------------------
-def test_oob_for_issue_29():
+
+
+class TestIssue29:
     dta = sm.datasets.sunspots.load_pandas().data
     dta.index = pd.Index(sm.tsa.datetools.dates_from_range('1700', '2008'))
     del dta["YEAR"]
 
     xreg = np.random.RandomState(1).rand(dta.shape[0], 3)
 
-    # Try for cv on/off, various D levels, and various Xregs
-    for d in (0, 1):
-        for cv in (0, 3):
-            for exog in (xreg, None):
+    @pytest.mark.parametrize('d', [0, 1])
+    @pytest.mark.parametrize('cv', [0, 3])
+    @pytest.mark.parametrize('exog', [xreg, None])
+    def test_oob_for_issue_29(self, d, cv, exog):
+        model = ARIMA(order=(2, d, 0),
+                      out_of_sample_size=cv).fit(self.dta,
+                                                 exogenous=exog)
 
-                # surround with try/except so we can log the failing combo
-                try:
-                    model = ARIMA(order=(2, d, 0),
-                                  out_of_sample_size=cv).fit(dta,
-                                                             exogenous=exog)
+        # If exogenous is defined, we need to pass n_periods of
+        # exogenous rows to the predict function. Otherwise we'll
+        # just leave it at None
+        if exog is not None:
+            xr = exog[:3, :]
+        else:
+            xr = None
 
-                    # If exogenous is defined, we need to pass n_periods of
-                    # exogenous rows to the predict function. Otherwise we'll
-                    # just leave it at None
-                    if exog is not None:
-                        xr = exog[:3, :]
-                    else:
-                        xr = None
-
-                    _, _ = model.predict(n_periods=3, return_conf_int=True,
-                                         exogenous=xr)
-
-                # Statsmodels can be fragile with ARMA coefficient
-                # computation. If we encounter that, pass:
-                #   ValueError: The computed initial MA coefficients are
-                #       not invertible. You should induce invertibility,
-                #       choose a different model order, or ...
-                except Exception as ex:
-                    # print("Failing combo: d=%i, cv=%i, exog=%r"
-                    #       % (d, cv, exog))
-                    if "invertibility" in pytest_error_str(ex):
-                        pass
-                    else:
-                        raise
-
-
-def test_oob_with_zero_out_of_sample_size():
-    with pytest.warns(UserWarning) as uw:
-        auto_arima(y, suppress_warnings=False, information_criterion="oob",
-                   out_of_sample_size=0)
-
-    assert uw[0].message.args[0] == "information_criterion cannot be 'oob' " \
-                                    "with out_of_sample_size = 0. Falling " \
-                                    "back to information criterion = aic."
-
-
-def test_issue_30():
-    # From the issue:
-    vec = np.array([33., 44., 58., 49., 46., 98., 97.])
-
-    arm = AutoARIMA(out_of_sample_size=1, seasonal=False,
-                    suppress_warnings=True)
-    arm.fit(vec)
-
-    # This is a way to force it:
-    ARIMA(order=(0, 1, 0), out_of_sample_size=1).fit(vec)
-
-    # Want to make sure it works with exog arrays as well
-    exog = np.random.RandomState(1).rand(vec.shape[0], 2)
-    auto_arima(vec, exogenous=exog, out_of_sample_size=1,
-               seasonal=False,
-               suppress_warnings=True)
-
-    # This is a way to force it:
-    ARIMA(order=(0, 1, 0), out_of_sample_size=1).fit(vec, exogenous=exog)
+        _, _ = model.predict(n_periods=3, return_conf_int=True,
+                             exogenous=xr)
 
 
 def _try_get_attrs(arima):
@@ -405,27 +366,26 @@ def test_more_elaborate():
         y=hr, exogenous=xreg)
     _try_get_attrs(arima)
 
-    # pickle this for the __get/setattr__ coverage.
-    # since the only time this is tested is in parallel in auto.py,
-    # this doesn't actually get any coverage proof...
-    fl = 'some_temp_file.pkl'
-    with open(fl, 'wb') as p:
-        pickle.dump(arima, p)
+    with tempfile.TemporaryDirectory() as tdir:
 
-    # show we can predict with this even though it's been pickled
-    new_xreg = rs.rand(5, 4)
-    _preds = arima.predict(n_periods=5, exogenous=new_xreg)
+        # pickle this for the __get/setattr__ coverage.
+        # since the only time this is tested is in parallel in auto.py,
+        # this doesn't actually get any coverage proof...
+        fl = os.path.join(tdir, 'some_temp_file.pkl')
+        with open(fl, 'wb') as p:
+            pickle.dump(arima, p)
 
-    # now unpickle
-    with open(fl, 'rb') as p:
-        other = pickle.load(p)
+        # show we can predict with this even though it's been pickled
+        new_xreg = rs.rand(5, 4)
+        _preds = arima.predict(n_periods=5, exogenous=new_xreg)
 
-    # show we can still predict, compare
-    _other_preds = other.predict(n_periods=5, exogenous=new_xreg)
-    assert_array_almost_equal(_preds, _other_preds)
+        # now unpickle
+        with open(fl, 'rb') as p:
+            other = pickle.load(p)
 
-    # now remove the pickle file
-    os.unlink(fl)
+        # show we can still predict, compare
+        _other_preds = other.predict(n_periods=5, exogenous=new_xreg)
+        assert_array_almost_equal(_preds, _other_preds)
 
     # now show that since we fit the ARIMA with an exogenous array,
     # we need to provide one for predictions otherwise it breaks.
@@ -470,71 +430,7 @@ def test_the_r_src():
     assert abs(135.28 - fit.aic()) < 1.0  # R's is 135.28
 
 
-@pytest.mark.parametrize(
-    'endog, kwargs', [
-        # show we fail for bad start/max p, q values:
-        pytest.param(abc, {'start_p': -1}),
-        pytest.param(abc, {'start_q': -1}),
-        pytest.param(abc, {'max_p': -1}),
-        pytest.param(abc, {'max_q': -1}),
-
-        # show we fail when start < max:
-        pytest.param(abc, {'start_p': 1, 'max_p': 0}),
-        pytest.param(abc, {'start_q': 1, 'max_q': 0}),
-
-        # other assertions
-        pytest.param(abc, {'max_order': -1, 'stepwise': False}),
-        pytest.param(abc, {'max_d': -1}),
-        pytest.param(abc, {'d': -1}),
-        pytest.param(abc, {'max_D': -1}),
-        pytest.param(abc, {'D': -1}),
-        pytest.param(abc, {'information_criterion': 'bad-value'}),
-        pytest.param(abc, {'m': 0}),
-    ]
-)
-def test_value_errors(endog, kwargs):
-    with pytest.raises(ValueError):
-        auto_arima(endog, **kwargs)
-
-
-@pytest.mark.parametrize(
-    'endog, max_order, kwargs', [
-        # show that for starting values > max_order, we can still get a fit
-        pytest.param(abc, 3, {'start_p': 5,
-                              'start_q': 5,
-                              'seasonal': False,
-                              'stepwise': False}),
-
-        pytest.param(abc, 3, {'start_p': 5,
-                              'start_q': 5,
-                              'start_P': 2,
-                              'start_Q': 2,
-                              'seasonal': True,
-                              'stepwise': False}),
-    ]
-)
-def test_valid_max_order_edges(endog, max_order, kwargs):
-    fit = auto_arima(endog, max_order=max_order, **kwargs)
-    order = fit.order
-    ssnal = fit.seasonal_order
-    assert (sum(order) + sum(ssnal[:3])) <= max_order
-
-
-def test_many_orders():
-    lam = 0.5
-    lynx_bc = ((lynx ** lam) - 1) / lam
-    auto_arima(lynx_bc, start_p=1, start_q=1, d=0, max_p=5, max_q=5,
-               suppress_warnings=True, stepwise=True)
-
-
-def test_small_samples():
-    # if n_samples < 10, test the new starting p, d, Q
-    samp = lynx[:8]
-    auto_arima(samp, suppress_warnings=True, stepwise=True,
-               error_action='ignore')
-
-
-def test_with_seasonality1():
+def test_with_seasonality():
     fit = ARIMA(order=(1, 1, 1),
                 seasonal_order=(0, 1, 1, 12),
                 suppress_warnings=True).fit(y=wineind)
@@ -556,171 +452,18 @@ def test_with_seasonality1():
     fit.predict(n_periods=10, return_conf_int=True, alpha=0.05)
 
 
-def test_with_seasonality2():
-    def do_fit(simple_differencing=False):
-        return auto_arima(wineind, start_p=1, start_q=1, max_p=2,
-                          max_q=2, m=2, start_P=0,
-                          seasonal=True, n_jobs=2,
-                          d=1, D=1, stepwise=False,
-                          suppress_warnings=True,
-                          error_action='ignore',
-                          n_fits=20, random_state=42,
-                          sarimax_kwargs={
-                              'simple_differencing': simple_differencing},
-
-                          # Set to super low iter to make test move quickly
-                          max_order=None,
-                          maxiter=2)
-
-    # show that we can forecast even after the
-    # pickling (this was fit in parallel)
-    seasonal_fit = do_fit()
-    seasonal_fit.predict(n_periods=10)
-
-    # ensure summary still works
-    seasonal_fit.summary()
-
-    # Show we can predict on seasonal where conf_int is true
-    seasonal_fit.predict(n_periods=10, return_conf_int=True)
-
-    # We should get the same order when simple_differencing
-    simple = do_fit(True)
-    assert simple.order == seasonal_fit.order
-    assert simple.seasonal_order == seasonal_fit.seasonal_order
-
-
-def test_with_seasonality3():
-    # show we can estimate D even when it's not there...
-    auto_arima(wineind, start_p=1, start_q=1, max_p=2, max_q=2, m=wineind_m,
-               start_P=0, seasonal=True, d=1, D=None,
-               error_action='ignore', suppress_warnings=True,
-               trace=True,  # get the coverage on trace
-               random_state=42, stepwise=True,
-
-               # Set to super low iter to make test move quickly
-               maxiter=5)
-
-
-def test_with_seasonality4():
-    # show we can run a random search much faster! and while we're at it,
-    # make the function return all the values. Also, use small M to make our
-    # lives easier.
-    auto_arima(wineind, start_p=1, start_q=1, max_p=2, max_q=2, m=12,
-               start_P=0, seasonal=True, n_jobs=1, d=1, D=None, stepwise=False,
-               error_action='ignore', suppress_warnings=True,
-               random=True, random_state=42, return_valid_fits=True,
-               n_fits=3,  # only a few
-
-               # Set to super low iter to make test move quickly
-               maxiter=5)
-
-
-def test_with_seasonality5():
-    # can we fit the same thing with an exogenous array of predictors?
-    # also make it stationary and make sure that works...
-    # 9/22/18 - make not parallel to reduce mem overhead on pytest
-    all_res = auto_arima(wineind, start_p=1, start_q=1, max_p=2,
-                         max_q=2, m=12, start_P=0, seasonal=True,
-                         d=1, D=None, error_action='ignore',
-                         suppress_warnings=True, stationary=True,
-                         random_state=42, return_valid_fits=True,
-                         stepwise=True,
-                         exogenous=rs.rand(wineind.shape[0], 4),  # only fit 2
-
-                         # Set to super low iter to make test move quickly
-                         maxiter=5)
-
-    # show it is a list
-    assert hasattr(all_res, '__iter__')
-
-
-def test_with_seasonality6():
-    # show that we can fit an ARIMA where the max_p|q == start_p|q
-    auto_arima(hr, start_p=0, max_p=0, d=0, start_q=0, max_q=0,
-               seasonal=False, max_order=np.inf,
-               suppress_warnings=True)
-
-    # FIXME: we get an IndexError from statsmodels summary if (0, 0, 0)
-
-
-@pytest.mark.parametrize('as_series', [True, False])
-def test_with_seasonality7(as_series):
-    endog = wineind
-    if as_series:
-        endog = pd.Series(wineind)
-    # show we can fit one with OOB as the criterion
-    auto_arima(endog, start_p=1, start_q=1, max_p=2, max_q=2, m=12,
-               start_P=0, seasonal=True, n_jobs=1, d=1, D=1,
-               out_of_sample_size=10, information_criterion='oob',
-               suppress_warnings=True,
-               error_action='raise',  # do raise so it fails fast
-               random=True, random_state=42, n_fits=2,
-               stepwise=False,
-
-               # Set to super low iter to make test move quickly
-               maxiter=5)
-
-
-def test_corner_cases():
-    with pytest.raises(ValueError):
-        auto_arima(wineind, error_action='some-bad-string')
-
-    # things that produce warnings
-    with pytest.warns(UserWarning):
-        # show a constant result will result in a quick fit
-        auto_arima(np.ones(10), suppress_warnings=True)
-
-        # show the same thing with return_all results in the ARIMA in a list
-        fits = auto_arima(np.ones(10), suppress_warnings=True,
-                          return_valid_fits=True)
-        assert hasattr(fits, '__iter__')
-
-    # show we fail for n_iter < 0
-    with pytest.raises(ValueError):
-        auto_arima(np.ones(10), random=True, n_fits=-1)
-
-    # show if max* < start* it breaks:
-    with pytest.raises(ValueError):
-        auto_arima(np.ones(10), start_p=5, max_p=0)
-
-
-@pytest.mark.parametrize(
-    'data,test,m,expected', [
-        pytest.param(wineind, 'ch', 52, 2),
-        pytest.param(wineind, 'ch', 12, 0),
-        pytest.param(wineind, 'ocsb', 52, 0),
-        pytest.param(austres, 'ocsb', 4, 0)
-    ]
-)
-def test_nsdiffs_on_various(data, test, m, expected):
-    assert nsdiffs(data, m=m, test=test, max_D=3) == expected
-
-
-# Asserting where D grows too large as a product of an M that's too big.
-def test_m_too_large():
-    train = lynx[:90]
-
-    with pytest.raises(ValueError) as v:
-        auto_arima(train, start_p=1, start_q=1, start_P=1, start_Q=1,
-                   max_p=5, max_q=5, max_P=5, max_Q=5, seasonal=True,
-                   stepwise=True, suppress_warnings=True, D=10, max_D=10,
-                   error_action='ignore', m=20)
-
-    msg = pytest_error_str(v)
-    assert 'The seasonal differencing order' in msg
-
-
 # Test that (as of v0.9.1) we can pickle a model, pickle it again, load both
 # and create predictions.
 def test_double_pickle():
     arima = ARIMA(order=(0, 0, 0), trend='c', suppress_warnings=True)
     arima.fit(y)
 
-    # Now save it twice
-    file_a = 'first.pkl'
-    file_b = 'second.pkl'
+    with tempfile.TemporaryDirectory() as tdir:
 
-    try:
+        # Now save it twice
+        file_a = os.path.join(tdir, 'first.pkl')
+        file_b = os.path.join(tdir, 'second.pkl')
+
         # No compression
         joblib.dump(arima, file_a)
 
@@ -738,71 +481,6 @@ def test_double_pickle():
         pred_a = loaded_a.predict(n_periods=5)
         pred_b = loaded_b.predict(n_periods=5)
         assert np.allclose(pred_a, pred_b)
-
-    # Always remove in case we fail in try, leaving residual files
-    finally:
-        os.unlink(file_a)
-        os.unlink(file_b)
-
-
-def test_warn_for_large_differences():
-    # First: d is too large
-    with pytest.warns(ModelFitWarning):
-        auto_arima(wineind, seasonal=True, m=1, suppress_warnings=False,
-                   d=3, error_action='warn', maxiter=5)
-
-    # Second: D is too large. M needs to be > 1 or D will be set to 0...
-    # unfortunately, this takes a long time.
-    with pytest.warns(ModelFitWarning):
-        auto_arima(wineind, seasonal=True, m=2,  # noqa: F841
-                   suppress_warnings=False,
-                   D=3, error_action='warn',
-                   maxiter=5)
-
-
-def test_warn_for_stepwise_and_parallel():
-    with pytest.warns(UserWarning):
-        auto_arima(lynx, suppress_warnings=False, d=1,  # noqa: F841
-                   error_action='ignore', stepwise=True, n_jobs=2)
-
-
-# Force case where data is simple polynomial after differencing
-@pytest.mark.filterwarnings('ignore:divide by zero')  # Expected, so ignore
-def test_force_polynomial_error():
-    x = np.array([1, 2, 3, 4, 5, 6, 7, 8])
-    d = 3
-    xreg = None
-
-    with pytest.raises(ValueError) as ve, \
-            pytest.warns(ModelFitWarning) as mfw:
-        auto_arima(x, d=d, D=0, seasonal=False, exogenous=xreg, trace=2)
-
-    err_msg = pytest_error_str(ve)
-    assert 'simple polynomial' in err_msg, err_msg
-
-    warning_msgs = pytest_warning_messages(mfw)
-    assert any('more differencing operation' in w for w in warning_msgs)
-
-
-# Test if exogenous is not None and D > 0
-def test_seasonal_xreg_differencing():
-    # Test both a small M and a large M since M is used as the lag parameter
-    # in the xreg array differencing. If M is 1, D is set to 0
-    for m in (2,):  # 12): takes FOREVER
-        _ = auto_arima(wineind, d=1, D=1,  # noqa: F841
-                       seasonal=True,
-                       exogenous=wineind_xreg, error_action='ignore',
-                       suppress_warnings=True, m=m,
-
-                       # Set to super low iter to make test move quickly
-                       maxiter=5)
-
-
-# Show that we can complete when max order is None
-def test_inf_max_order():
-    _ = auto_arima(lynx, max_order=None,  # noqa: F841
-                   suppress_warnings=True,
-                   error_action='trace')
 
 
 # Regression testing for unpickling an ARIMA from an older version
@@ -834,9 +512,9 @@ def test_for_older_version():
         elif case == 3:
             arima.pkg_version_ = '0.0.1'  # will always be < than current
 
-        # Pickle it
-        pickle_file = 'model.pkl'
-        try:
+        with tempfile.TemporaryDirectory() as tdir:
+
+            pickle_file = os.path.join(tdir, 'model.pkl')
             joblib.dump(arima, pickle_file)
 
             # Now unpickle it and show that we get a warning (if expected)
@@ -849,9 +527,6 @@ def test_for_older_version():
             # we can still produce predictions (only if we fit)
             if do_fit:
                 arm.predict(n_periods=4)
-
-        finally:
-            os.unlink(pickle_file)
 
 
 @pytest.mark.parametrize(
@@ -1035,32 +710,3 @@ def test_update_1_iter(model):
 
     # They should be close
     assert np.allclose(params1, params2, atol=0.05)
-
-
-def test_AutoARIMA_class():
-    train, test = wineind[:125], wineind[125:]
-    mod = AutoARIMA(maxiter=5)
-    mod.fit(train)
-
-    endog = mod.model_.arima_res_.data.endog
-    assert_array_almost_equal(train, endog)
-
-    # update
-    mod.update(test, maxiter=2)
-    new_endog = mod.model_.arima_res_.data.endog
-    assert_array_almost_equal(wineind, new_endog)
-
-
-# "ValueError: negative dimensions are not allowed" in OCSB test
-def test_issue_191():
-    X = pd.read_csv(
-        os.path.join(abspath(dirname(__file__)), 'data', 'issue_191.csv'))
-    y = X[X.columns[1]].values
-    auto_arima(
-        y,
-        error_action="warn",
-        seasonal=True,
-        m=12,
-        alpha=0.05,
-        suppress_warnings=True,
-        trace=True)
